refactor(product): http 진입점을 api/http 패키지로 이동
