+++ service/product-service/src/main/java/dev/labs/commerce/product/api/http/dto/ModifyProductRequest.java
@@ -1,6 +1,6 @@
-package dev.labs.commerce.product.api.dto;
+package dev.labs.commerce.product.api.http.dto;
 
-public record RegisterProductRequest(
+public record ModifyProductRequest(
         String productName,
         Long price,
         String currency,
